refactor: OverrideAbleComponentFunctionComponent to type augment
