+++ .changeset/rich-snails-happen.md
@@ -1,0 +1,5 @@
+---
+'web': patch
+---
+
+Refactored OverrideAbleComponentFunctionComponent into a type augment
