+++ apps/web/src/components/container.tsx
@@ -1,13 +1,12 @@
 import clsx from 'clsx';
 import { PropsWithChildren } from 'react';
-import { OverrideAbleComponentFunctionComponent } from './override-able-component-function-component';
 
 interface Props extends PropsWithChildren {
   bleed?: boolean;
   className?: string;
 }
 
-export const Container: OverrideAbleComponentFunctionComponent<'div', Props> = ({
+export const Container: React.OverrideAbleComponentFC<'div', Props> = ({
   component: Component = 'div',
   bleed,
   children,
